requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
# PBR should always appear first
pbr!=2.1.0,>=2.0.0 # Apache-2.0
django-file-md5
python-qinlingclient>=1.1.0 # Apache-2.0

# This will be installed from git in OpenStack CI if the job setting
# required-projects for horizon:
horizon>=14.0.0.0b1 # Apache-2.0
